updates: check at the beginning if previous build dir or build.Xml exists, if yes delete them

+++ utils/part1000/build_cr.sh
@@ -12,24 +12,28 @@ exec 2>&1
 
 if [ -z $1 ] || [ -z $2 ];
 then
-    echo "Error: first, second or both arguments missing. Two arguments to be specified: 1) absolute path ending in 'stepmod', with no trailing slash, of checked-out stepmod repository (e.g.: /Users/klt/Projets/workspace/stepmod); 2) name of the CR (e.g.: CR_PDM_1)."
+echo "Error: first, second or both arguments missing. Two arguments to be specified: 1) absolute path ending in 'stepmod', with no trailing slash, of checked-out stepmod repository (e.g.: /Users/klt/Projets/workspace/stepmod); 2) name of the CR (e.g.: CR_PDM_1)."
 exit
 fi
 
+echo "============================ Deleting eventual provious build directory and build.xml: ============================"
+
+if [ -d $1/publication/isopub/$2 ];
+then
+rm -rfv stepmod/publication/isopub/$2
+fi
+
 cd $1/publication/part1000/$2/
+
+if [ -f build.xml ];
+then
+rm -rfv build.xml
 
 ant -lib $1/etc/saxon6-5-5/saxon.jar -lib $1/etc/saxon6-5-5/saxon-xml-apis.jar -lib $1/etc/saxon6-5-5/saxon-jdom.jar -f buildbuild.xml
 
 ant -lib $1/etc/saxon6-5-5/saxon.jar -lib $1/etc/saxon6-5-5/saxon-xml-apis.jar -lib $1/etc/saxon6-5-5/saxon-jdom.jar all
 
-cd $1/..
-
-echo "============================ Copying build directory from stepmod/publication/isopub to workspace: ============================"
-
-cp -Riv stepmod/publication/isopub/$2 .
-
-echo "============================ Deleting build directory from stepmod/publication/isopub: ============================"
-
-rm -rfv stepmod/publication/isopub/$2
+fi
 
 exit
+
